+++ android/app/src/main/java/com/created/team201/presentation/studyList/StudyListFragment.kt
@@ -10,6 +10,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.created.team201.R
 import com.created.team201.databinding.FragmentStudyListBinding
 import com.created.team201.presentation.common.BindingFragment
+import com.created.team201.presentation.studyList.adapter.StudyListAdapter
 
 class StudyListFragment : BindingFragment<FragmentStudyListBinding>(R.layout.fragment_study_list) {
