+++ android/app/src/main/java/com/created/team201/presentation/studyList/adapter/StudyListAdapter.kt
@@ -1,8 +1,10 @@
-package com.created.team201.presentation.studyList
+package com.created.team201.presentation.studyList.adapter
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
+import com.created.team201.presentation.studyList.model.StudySummaryBackground
+import com.created.team201.presentation.studyList.model.StudySummaryUiModel
 
 class StudyListAdapter : ListAdapter<StudySummaryUiModel, StudyListViewHolder>(StudyListDiffUtil) {
