+++ android/app/src/main/java/com/created/team201/presentation/studyList/adapter/StudyListViewHolder.kt
@@ -1,4 +1,4 @@
-package com.created.team201.presentation.studyList
+package com.created.team201.presentation.studyList.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,6 +6,7 @@ import androidx.core.content.ContextCompat.getDrawable
 import androidx.recyclerview.widget.RecyclerView
 import com.created.team201.R
 import com.created.team201.databinding.ItemStudyListBinding
+import com.created.team201.presentation.studyList.model.StudySummaryUiModel
 
 class StudyListViewHolder(private val parent: ViewGroup) : RecyclerView.ViewHolder(
     LayoutInflater.from(parent.context).inflate(R.layout.item_study_list, parent, false),
