+++ android/app/src/main/java/com/created/team201/presentation/studyList/model/StudyStatusImage.kt
@@ -1,4 +1,4 @@
-package com.created.team201.presentation.studyList
+package com.created.team201.presentation.studyList.model
 
 import com.created.team201.R
